diff shrink
