+++ nexusd/src/trust/mod.rs
@@ -1,7 +1,7 @@
 mod engine;
 mod export;
 mod job;
-mod neo4j;
+pub mod neo4j;
 
 pub use engine::{TrustRankEngine, TrustRankParams, TrustRankStats};
 pub use export::read_scores;
